Save filter state+keywords using localStorage
This commit modifies the tampermonkey script to save the filter state and the excluded keywords using localStorage. This way, the variables do not reset when the page is refreshed.

If there is nothing saved to the localStorage, the script will use the default values for the filterOn and excludeJobTitle variables. The filterOn variable will be set to false, meaning that the filter will be off by default. The excludeJobTitle variable will be set to an empty array, meaning that no keywords will be excluded by default. The inputBox element will also have an empty value, indicating that the user can enter any keywords they want to exclude from the job listings.