Refactor to hide parent article of hidden jobs
This commit refactors the code that hides the jobs that do not meet the criteria of quick apply, good badge and excluded title. Previously, the code only hid the job element itself, but not the parent article element that contains it. This resulted in empty spaces in the page layout. Now, the code also checks if the parent element is an article and hides it accordingly. This improves the user experience and the page aesthetics.